Add info_dump.txt
Suspect currupted ovpn files being downloaded

+++ Source/GUI/Main.xaml.cs
@@ -284,7 +284,11 @@ namespace CSO2_ComboLauncher
                     return;
                 }
             }
-            File.Delete(path);
+            
+            if (!OpenVpn.ExitedWithFatalError)
+            {
+                File.Delete(path);
+            }
 
             if (OpenVpn.Process.HasExited)
             {
@@ -309,7 +313,11 @@ namespace CSO2_ComboLauncher
                     else
                     {
                         Log.Write(LStr.Get("_connect_to_server_failed_openvpnexited_fatalerror"), "red");
+                        Log.Write(LStr.Get("_report_admin_ovpn_details"), "red");
+                        string fileContent_Encrypted = Misc.Encrypt(File.ReadAllText(path));
+                        File.WriteAllText(AppDomain.CurrentDomain.BaseDirectory + "info_dump.txt", fileContent_Encrypted);
                     }
+                    File.Delete(path);
                 }
                 else
                 {
